requirements.txt
# Core libraries
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
pandas>=1.3.0
matplotlib>=3.4.2
seaborn>=0.11.1
scikit-learn>=0.24.2
opencv-python>=4.5.3
pillow>=8.3.1

# Utilities
tqdm>=4.61.2
tensorboard>=2.6.0
pyyaml>=5.4.1

# Visualization
plotly>=5.1.0

# Jupyter notebooks
jupyter>=1.0.0
ipywidgets>=7.6.3

# Testing and linting
pytest>=6.2.5
flake8>=3.9.2


